backend/requirements.txt
Flask==2.3.2
PyJWT==2.8.0
filelock==3.12.2
Flask-Cors==4.0.0
python-dotenv
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
Flask-Migrate==4.0.4
cryptography